logger: move Collector
